+++ setup.cfg
@@ -44,3 +44,7 @@ where=src
 [options.extras_require]
 dev =
     pre-commit
+[options.extras_require]
+test=
+  pytest-external-blockers
+  pytest-manual-marker
